requirements.txt
loguru==0.6.0
accelerate==0.21.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
attrs==22.2.0
bitsandbytes==0.39.1
cachelib==0.9.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cmake==3.25.0
cryptography==40.0.1
Cython==0.29.34
dataclasses-json==0.5.7
datasets==2.12.0
decorator==5.1.1
deepspeed==0.10.0
Deprecated==1.2.13
easydict==1.10
evaluate==0.4.0
fastapi==0.95.1
ffmpeg-python==0.2.0
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
h5py==3.8.0
huggingface-hub==0.14.1
jsonlines==3.1.0
Jinja2==3.1.2
#langchain==0.0.162
loguru==0.6.0
lxml==4.9.2
multiprocess==0.70.12.2
ninja==1.11.1
pandas==1.5.3
peft==0.4.0
protobuf==3.20.3
pycocotools==2.0.6
pydantic==1.9.0
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==5.3.1
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
streamlit==1.24.0
streamlit-chat==0.0.2.2
streamlit-extras
six==1.16.0
tqdm==4.65.0
transformers==4.31.0
trl==0.4.7
ujson==5.7.0
urllib3==1.26.5
uvicorn==0.21.1
Werkzeug==2.2.3
yacs==0.1.8
transformers_stream_generator==0.0.4
dynaconf==3.2.0
xlrd==2.0.1
xlwt==1.3.0
tensorboard==2.13.0
